+++ modules/kubernetes/aks-core/modules.tf
@@ -326,10 +326,8 @@ module "datadog" {
   namespace_include = compact(concat(var.namespaces[*].name, var.datadog_config.extra_namespaces))
 }
 
-# grafana-agent
 module "grafana_agent" {
   depends_on = [module.opa_gatekeeper]
-
   for_each = {
     for s in ["grafana-agent"] :
     s => s
@@ -338,25 +336,51 @@ module "grafana_agent" {
 
   source = "../../kubernetes/grafana-agent"
 
+  cluster_name    = "${var.name}${var.aks_name_suffix}"
+  environment     = var.environment
+  cloud_provider  = "azure"
+  remote_logs_url = var.grafana_agent_config.remote_logs_url
+  azure_config = {
+    azure_key_vault_name = var.grafana_agent_config.azure_key_vault_name
+    identity = {
+      client_id   = var.grafana_agent_config.identity.client_id
+      resource_id = var.grafana_agent_config.identity.resource_id
+      tenant_id   = var.grafana_agent_config.identity.tenant_id
+    }
+  }
+}
+
+# grafana-agent
+module "grafana_agent_tenant" {
+  depends_on = [module.grafana_agent]
+
+  for_each = {
+    for s in ["grafana-agent-tenant"] :
+    s => s
+    if var.grafana_agent_tenant_enabled
+  }
+
+  source = "../../kubernetes/grafana-agent-tenant"
+
   remote_write_urls = {
-    metrics = var.grafana_agent_config.remote_write_urls.metrics
-    logs    = var.grafana_agent_config.remote_write_urls.logs
-    traces  = var.grafana_agent_config.remote_write_urls.traces
+    metrics = var.grafana_agent_tenant_config.remote_write_urls.metrics
+    logs    = var.grafana_agent_tenant_config.remote_write_urls.logs
+    traces  = var.grafana_agent_tenant_config.remote_write_urls.traces
   }
 
   credentials = {
-    metrics_username = var.grafana_agent_config.credentials.metrics_username
-    metrics_password = var.grafana_agent_config.credentials.metrics_password
-    logs_username    = var.grafana_agent_config.credentials.logs_username
-    logs_password    = var.grafana_agent_config.credentials.logs_password
-    traces_username  = var.grafana_agent_config.credentials.traces_username
-    traces_password  = var.grafana_agent_config.credentials.traces_password
+    metrics_username = var.grafana_agent_tenant_config.credentials.metrics_username
+    metrics_password = var.grafana_agent_tenant_config.credentials.metrics_password
+    logs_username    = var.grafana_agent_tenant_config.credentials.logs_username
+    logs_password    = var.grafana_agent_tenant_config.credentials.logs_password
+    traces_username  = var.grafana_agent_tenant_config.credentials.traces_username
+    traces_password  = var.grafana_agent_tenant_config.credentials.traces_password
   }
 
   cluster_name      = "${var.name}${var.aks_name_suffix}"
   environment       = var.environment
   vpa_enabled       = var.vpa_enabled
-  namespace_include = compact(concat(var.namespaces[*].name, var.grafana_agent_config.extra_namespaces))
+  namespace_include = compact(concat(var.namespaces[*].name, var.grafana_agent_tenant_config.extra_namespaces))
 }
 
 # falco