+++ modules/kubernetes/aks-core/networkpolicy.tf
@@ -45,7 +45,7 @@ resource "kubernetes_network_policy" "allow_egress_ingress_grafana_agent" {
   for_each = {
     for ns in var.namespaces :
     ns.name => ns
-    if var.grafana_agent_enabled && var.kubernetes_network_policy_default_deny
+    if var.grafana_agent_tenant_enabled && var.kubernetes_network_policy_default_deny
   }
 
   metadata {