+++ validation/kubernetes/grafana-agent/main.tf
@@ -7,23 +7,8 @@ provider "helm" {}
 module "grafana_agent" {
   source = "../../../modules/kubernetes/grafana-agent"
 
-  remote_write_urls = {
-    metrics = "foo"
-    logs    = "bar"
-    traces  = "baz"
-  }
-
-  credentials = {
-    metrics_username = "foo"
-    metrics_password = "bar"
-    logs_username    = "foo"
-    logs_password    = "bar"
-    traces_username  = "foo"
-    traces_password  = "bar"
-  }
-
-  cluster_name      = "aks1"
-  environment       = "dev"
-  vpa_enabled       = false
-  namespace_include = ["foobar"]
+  cluster_name    = "aks1"
+  environment     = "dev"
+  cloud_provider  = "azure"
+  remote_logs_url = "http://unbox-loki-distributed-distributor.monitor.svc.cluster.local:3100/loki/api/v1/push"
 }