+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-### Fix
-
 - [#606](https://github.com/XenitAB/terraform-modules/pull/606) Fix electionID on ingress-nginx when using private and public ingress-nginx.
+
+### Added
+
+- [#602](https://github.com/XenitAB/terraform-modules/pull/602) Add platform loggin through Grafana Agent
 
 ## 2022.03.3